Support inline credentials and propagationDelay in config
If `config.credentials` is passed as a string, it will continue to be
interpreted as a filename; otherwise it will be interpreted as a JS object. This
provides more flexibility to configure the module in certain contexts.

Also supports a configurable propagationDelay, which defaults to 5000
ms. Previously, with no propagationDelay, acme.js would print this warning:

  warning Error: dns-01 challenge's `propagationDelay` not set, defaulting to 5000ms
  code: 'E_NO_DNS_DELAY',
  description: "Each dns-01 challenge should specify
    challenges['dns-01'].propagationDelay as an estimate of how long DNS
    propagation will take."

In practice, I have found that the default of 5sec was not long enough to
reliably succeed. At 30sec I saw some failures but mostly success. I suggest
using 40sec.